backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
alembic==1.14.1
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==5.2.1
celery==5.4.0
slowapi==0.1.9
bleach==6.2.0
cryptography==44.0.1
httpx==0.28.1
cloudinary==1.42.2

